OpenPDFCommand と ResetPreviewsCommand を OpenPDFCommand に統合し、UIスレッドで重い処理を実行しないように気を付けるようにタスクを書き直した
